Monthly and daily preview of tasks
Added Task model class. Updated the fragment for the calendar (monthly preview) using the MaterialCalendarView. Added a decorator for marking events on the calendar. Created a fragment for the daily preview of tasks using RecyclerView.  Added the adapter for the RecyclerView Updated build.gradle with dependencies for the MaterialCalendarView and RecyclerView.