Task labels
Updated the database with a table for labels and a join table for tasks and labels. Added the LabelContentProvider. Updated authorities because of multiple content providers. Updated task detail preview with task labels from the database. Implemented adding existing/creating new labels in adding a new task. Updated the dialog for labels with an autocomplete text view that shows existing labels.